Remove c implementation
I believe the python program does what it needs to do and is
plenty extensible for the future. In the worst case we can still write
c code and have it run in python.

Since there is no real-time or efficiency constraint in *kicking off*
the programs, the ease of development in python warrants using it.