Fix init script pin config file creation
The script was not running the command with sudo as expected because
Shell i/o redirection happens before the shell executes the command,
i.e. the `> /etc/hobby-hub/pin_mapping.json` was performed as the
current user, not root.

This is a workaround. See https://bit.ly/3sp6N8R